Just in case

+++ app/src/main/java/com/orka/myfinances/data/models/product/Product.kt
@@ -6,8 +6,8 @@ import com.orka.myfinances.data.models.folder.Warehouse
 data class Product(
     val id: Id,
     val name: String,
-    val price: Double,
-    val salePrice: Double,
+    val price: Int,
+    val salePrice: Int,
     val warehouse: Warehouse,
     val properties: List<Property>,
     val description: String = ""
